ISPN-4839 Add the CDI + JCache to the Embedded UberJar and CDI to the Remote UberJar
