+++ all/embedded-query/pom.xml
@@ -63,6 +63,12 @@
                         </excludes>
                      </artifactSet>
                      <createSourcesJar>true</createSourcesJar>
+                     <relocations>
+                        <relocation>
+                           <pattern>org.jboss.logging</pattern>
+                           <shadedPattern>infinispan.org.jboss.logging</shadedPattern>
+                        </relocation>
+                     </relocations>
                      <transformers>
                         <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
                            <manifestEntries>
