+++ cdi/pom.xml
@@ -39,10 +39,12 @@
       <dependency>
          <groupId>org.infinispan</groupId>
          <artifactId>infinispan-core</artifactId>
+         <optional>true</optional>
       </dependency>
       <dependency>
          <groupId>org.infinispan</groupId>
          <artifactId>infinispan-client-hotrod</artifactId>
+         <optional>true</optional>
       </dependency>
 
       <!-- Logging dependency -->
