+++ distribution/package.xml
@@ -32,9 +32,7 @@
       </copy>
 
       <!-- Modules -->
-      <module dir="${output.dir}" target="modules/cdi" module="cdi" artifact="infinispan-cdi" />
       <module dir="${output.dir}" target="modules/infinispan-extended-statistics" module="extended-statistics" artifact="infinispan-extended-statistics" />
-      <module dir="${output.dir}" target="modules/jcache" module="jcache" artifact="infinispan-jcache" />
       <module dir="${output.dir}" target="modules/persistence/remote" module="persistence/remote" artifact="infinispan-cachestore-remote" />
       <module dir="${output.dir}" target="modules/persistence/rest" module="persistence/rest" artifact="infinispan-cachestore-rest" />
       <module dir="${output.dir}" target="modules/spring" module="spring" artifact="infinispan-spring" />
@@ -70,6 +68,7 @@
          </copy>
          <mkdir dir="@{dir}/lib" />
          <copy todir="@{dir}/lib">
+            <fileset file="${javax.cache:cache-api:jar}" />
             <fileset file="${org.jboss.spec.javax.transaction:jboss-transaction-api_1.1_spec:jar}" />
             <mapper type="flatten" />
          </copy>
