+++ integrationtests/all-embedded-query-it/pom.xml
@@ -1,5 +1,7 @@
 <?xml version='1.0' encoding='UTF-8'?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+<project
+   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+   xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
    <modelVersion>4.0.0</modelVersion>
 
    <parent>
@@ -11,7 +13,7 @@
 
    <artifactId>infinispan-embedded-query-it</artifactId>
    <name>Infinispan Embedded Query Integration Tests</name>
-<description>Infinispan Embedded Query All-in-One module Integration Tests</description>
+   <description>Infinispan Embedded Query All-in-One module Integration Tests</description>
 
    <dependencies>
       <dependency>
@@ -25,9 +27,9 @@
          <scope>provided</scope>
       </dependency>
       <dependency>
-         <groupId>org.testng</groupId>
-         <artifactId>testng</artifactId>
-         <version>6.8</version>
+         <groupId>junit</groupId>
+         <artifactId>junit</artifactId>
+         <version>4.11</version>
          <scope>test</scope>
       </dependency>
    </dependencies>
